shanged structure files
