+++ src/main/java/LoginPageMesto.java
@@ -1,0 +1,30 @@
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+// Класс страницы авторизации(просто проверка, что можно зайтит на страницу)
+public class LoginPageMesto {
+
+    private WebDriver driver;
+    private By emailField = By.id("email");
+    private By passwordField = By.id("password");
+    private By signInButton = By.className("auth-form__button");
+
+    public LoginPageMesto(WebDriver driver){
+        this.driver = driver;
+    }
+
+    public void setUsername(String username) {
+        driver.findElement(emailField).sendKeys(username);
+    }
+    public void setPassword(String password) {
+        driver.findElement(passwordField).sendKeys(password);
+    }
+    public void clickSignInButton() {
+        driver.findElement(signInButton).click();
+    }
+    public void login(String username, String password){
+        setUsername(username);
+        setPassword(password);
+        clickSignInButton();
+    }
+}
