+++ src/main/java/MainPageScooter.java
@@ -1,0 +1,30 @@
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+
+public class MainPageScooter { //главная страница
+         //проверяем, что страница загружена
+         //кнопка заказа вписать сюда
+         private WebDriver driver;
+         private By homeHeader = By.className("Home_Header__iJKdX");//заголовок страницы
+         private By ButtonOrderUp = By.className("Button_Button__ra12g");
+         private By ButtonOderDown = By.className("Button_Middle__1CSJM");
+
+         public MainPageScooter(WebDriver driver){
+             this.driver = driver;
+         }
+         // метод ожидания загрузки страницы, проверяем, что тсраница загружена
+         public void waitForLoadHeader(){
+             new WebDriverWait(driver, Duration.ofSeconds(15))
+                     .until(ExpectedConditions.visibilityOfElementLocated(homeHeader));
+         }
+         // метод для получения текста элемента в заголовке
+         public String titleInHeader(){
+             return driver.findElement(homeHeader).getText();
+         }
+
+
+    }
