+++ src/test/java/Scooter.java
@@ -68,47 +68,4 @@ public class Scooter {
     }
 }
 
-     class MainPageScooter { //главная страница
-         //проверяем, что страница загружена
-         //кнопка заказа вписать сюда
-         private WebDriver driver;
-         private By homeHeader = By.className("Home_Header__iJKdX");//заголовок страницы
-         private By ButtonOrderUp = By.className("Button_Button__ra12g");
-         private By ButtonOderDown = By.className("Button_Middle__1CSJM");
 
-         public MainPageScooter(WebDriver driver){
-             this.driver = driver;
-         }
-         // метод ожидания загрузки страницы, проверяем, что тсраница загружена
-         public void waitForLoadHeader(){
-             new WebDriverWait(driver, Duration.ofSeconds(15))
-                     .until(ExpectedConditions.visibilityOfElementLocated(homeHeader));
-         }
-         // метод для получения текста элемента в заголовке
-         public String titleInHeader(){
-             return driver.findElement(homeHeader).getText();
-         }
-
-
-    }
-
-
-    class FAQ{ //класс по проверке ответов
-        private WebDriver driver;
-
-        public FAQ(WebDriver driver){this.driver = driver;}
-
-        public void clickOpenListButton(int numberRow) {
-
-            By listQuestions = By.xpath(".//div[@id='accordion__heading-"+numberRow+"']");
-            driver.findElement(listQuestions).click();
-        }
-
-            public String checkListAnswer(int numberRow){
-
-               By listOfAnswer = By.xpath(".//div[@id='accordion__panel-" + numberRow + "']");
-               String text = driver.findElement(listOfAnswer).getText();
-                return text;
-            }
-
-    }
